refactor(mbk): extract attribution lease->property helpers to a sibling
The approve-fix pushed attribution_service.py to 530 LOC, tripping the >500
growth guard. Extract the two DB helpers (_get_lease_signed_applicants,
_get_property_id_for_applicant) into attribution_helpers.py and import them
back (~488 LOC now). No behavior change: every function stays put, so all
callers and test patch targets are unchanged (the only namespace-bound
patched name, unit_of_work, is used solely by functions that remain).
